refactor: migrar transações financeiras para BigDecimal

+++ src/br/com/bancodigital/interfaces/IConta.java
@@ -2,9 +2,11 @@ package br.com.bancodigital.interfaces;
 
 import br.com.bancodigital.models.Conta;
 
+import java.math.BigDecimal;
+
 public interface IConta {
-    void sacar(double valor);
-    void depositar(double valor);
-    void transferir(double valor, Conta destino);
+    void sacar(BigDecimal valor);
+    void depositar(BigDecimal valor);
+    void transferir(BigDecimal valor, Conta destino);
     void imprimirExtrato();
 }
